Moved CreateCheckpointBench away from wide, fixed formatting.
