+++ rspace-bench/src/test/scala/coop/rchain/rspace/bench/CreateCheckpointBench.scala
@@ -1,10 +1,10 @@
-package coop.rchain.rspace.bench.wide
+package coop.rchain.rspace.bench
 import java.util.concurrent.TimeUnit
 
 import coop.rchain.catscontrib.TaskContrib._
-import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.rspace.history.TrieCache
 import monix.eval.Task
+import monix.execution.Scheduler
 import org.openjdk.jmh.annotations._
 import org.openjdk.jmh.infra.Blackhole
 
@@ -17,6 +17,7 @@ class CreateCheckpointBench {
   @Warmup(iterations = 1)
   @Measurement(iterations = 1)
   def createCheckpoint_old(bh: Blackhole, state: CreateCheckpointBenchState): Unit = {
+    TrieCache.useCache = false
     bh.consume(state.runtime.space.createCheckpoint())
   }
 
@@ -27,29 +28,22 @@ class CreateCheckpointBench {
   @Threads(1)
   @Warmup(iterations = 1)
   @Measurement(iterations = 1)
-  def createCheckpoint_new(bh: Blackhole, state: CreateCheckpointBenchStateVsCache): Unit = {
+  def createCheckpoint_new(bh: Blackhole, state: CreateCheckpointBenchState): Unit = {
+    TrieCache.useCache = true
     bh.consume(state.runtime.space.createCheckpoint())
   }
 }
 
 @State(Scope.Benchmark)
-class CreateCheckpointBenchStateVsCache extends CreateCheckpointBenchState {
-  override val useCache = true
-}
+class CreateCheckpointBenchState extends EvalBenchStateBase {
+  override val rhoScriptSource: String = "/rholang/loop-with-wks.rho"
+  implicit val scheduler: Scheduler    = monix.execution.Scheduler.Implicits.global
 
-@State(Scope.Benchmark)
-class CreateCheckpointBenchState extends WideBenchBaseState {
-  override val rhoSetupScriptPath: String = "/rholang/loop-with-wks.rho"
-  override val rhoScriptSource: String    = "/rholang/loop-with-wks.rho"
-
-  implicit val rand: Blake2b512Random        = Blake2b512Random(128)
-  var runReplayTask: Task[Vector[Throwable]] = Task.now(Vector.empty)
-
-  val useCache = false
-
+  @Setup
   override def doSetup(): Unit = {
-    TrieCache.useCache = this.useCache
     super.doSetup()
+
+    println("runTask")
     val runTask = createTest(term, runtime.reducer)
     assert(runTask.unsafeRunSync.isEmpty)
   }
